+++ editors/helix/queries/nexium/indents.scm
@@ -1,0 +1,17 @@
+; Nexium indentation for Helix
+[
+  (block)
+  (field_list)
+  (arguments)
+  (parameters)
+  (struct_literal)
+  (array_literal)
+  (match_expression)
+  (parenthesized_expression)
+] @indent
+
+[
+  "}"
+  ")"
+  "]"
+] @outdent