+++ editors/zed/languages/nexium/indents.scm
@@ -1,0 +1,14 @@
+[
+  (block)
+  (field_list)
+  (arguments)
+  (parameters)
+  (struct_literal)
+  (array_literal)
+  (match_expression)
+  (parenthesized_expression)
+] @indent
+
+(_ "{" "}" @end) @indent
+(_ "(" ")" @end) @indent
+(_ "[" "]" @end) @indent